feat: jwt의존성 추가

+++ meerkatai/src/main/java/com/capstone/meerkatai/anomalybehavior/repository/AnomalyBehaviorRepository.java
@@ -3,5 +3,5 @@ package com.capstone.meerkatai.anomalybehavior.repository;
 import com.capstone.meerkatai.anomalybehavior.entity.AnomalyBehavior;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnomalyBehaviorRepository extends JpaRepository<AnomalyBehavior, Integer> {
+public interface AnomalyBehaviorRepository extends JpaRepository<AnomalyBehavior, Long> {
 }
