+++ meerkatai/src/main/java/com/capstone/meerkatai/dashboard/repository/DashboardRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface DashboardRepository extends JpaRepository<Dashboard, Integer> {
-    List<Dashboard> findByUserUserId(Integer userId);
+public interface DashboardRepository extends JpaRepository<Dashboard, Long> {
+    List<Dashboard> findByUserUserId(Long userId);
 }
