+++ meerkatai/src/main/java/com/capstone/meerkatai/streamingvideo/repository/StreamingVideoRepository.java
@@ -1,13 +1,11 @@
 package com.capstone.meerkatai.streamingvideo.repository;
 
-//import com.capstone.meerkatai.streamingvideo.entity.StreamingVideo;
-//import org.springframework.data.jpa.repository.JpaRepository;
+import com.capstone.meerkatai.streamingvideo.entity.StreamingVideo;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-//import java.util.List;
+import java.util.List;
 
-/*
-public interface StreamingVideoRepository extends JpaRepository<StreamingVideo, Integer> {
-    List<StreamingVideo> findByUserUserId(Integer userId);
-    List<StreamingVideo> findByCctvCctvId(Integer cctvId);
+public interface StreamingVideoRepository extends JpaRepository<StreamingVideo, Long> {
+  List<StreamingVideo> findByUserUserId(Long userId);
+  List<StreamingVideo> findByCctvCctvId(Long cctvId);
 }
-*/
