+++ meerkatai/src/main/java/com/capstone/meerkatai/video/dto/VideoDetailsResponse.java
@@ -6,17 +6,16 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class VideoDetailsResponse {
-    private Integer video_id;
+    private Long video_id;
     private String file_path;
     private String thumbnail_path;
     private Integer duration;
     private Integer file_size;
     private Boolean video_status;
     private String created_at;
-    private Integer streaming_video_id;
+    private Long streaming_video_id;
     private Long cctv_id;
     private String cctv_name;
     private Long user_id;
     private String anomaly_type;
 }
-
